feat: 카드 MVC 기능구현/카카오페이 정기결제 구현중

+++ src/main/java/com/elice/nbbang/domain/payment/dto/PaymentDto.java
@@ -15,14 +15,14 @@ public class PaymentDto {
     private String partnerUserId;
     private String partnerOrderId;
     private String paymentType;
-    private Long amount;
+    private Integer amount;
     private String status;
     private String paymentCreatedAt;
     private String cid;
     private String paymentApprovedAt;
     private String refundDate;
     private String cardCompany;
-    private Long refundAmount;
+    private Integer refundAmount;
     private String sid;
 
     public static PaymentDto fromEntity(Payment payment) {
